requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.14
Pillow>=10.0
reportlab>=4.0
openpyxl>=3.1
python-decouple>=3.8
django-crispy-forms>=2.0
crispy-tailwind>=1.0
gunicorn>=21.2
